+++ packages/components/src/hooks/index.ts
@@ -2,6 +2,7 @@ export { default as UIProvider, Toast as ToastProps, useUI } from './UIProvider'
 export { useFadeEffect } from './useFadeEffect'
 export { useTrapFocus } from './useTrapFocus'
 export { useSearch } from './useSearch'
+export { useScrollDirection } from './useScrollDirection'
 export { useSlider } from './useSlider'
 export type {
   UseSliderArgs,
@@ -10,3 +11,5 @@ export type {
   SlideDirection,
 } from './useSlider'
 export { useSlideVisibility } from './useSlideVisibility'
+
+