+++ packages/components/src/molecules/NavbarSlider/index.ts
@@ -1,0 +1,11 @@
+export { default } from './NavbarSlider'
+export type { NavbarSliderProps } from './NavbarSlider'
+
+export { default as NavbarSliderHeader } from './NavbarSliderHeader'
+export type { NavbarSliderHeaderProps } from './NavbarSliderHeader'
+
+export { default as NavbarSliderContent } from './NavbarSliderContent'
+export type { NavbarSliderContentProps } from './NavbarSliderContent'
+
+export { default as NavbarSliderFooter } from './NavbarSliderFooter'
+export type { NavbarSliderFooterProps } from './NavbarSliderFooter'